操作数据路的后三个接口
