+++ app/db/operate_database.py
@@ -1,14 +1,5 @@
 from pymongo import MongoClient
-from pydantic import BaseModel
-
-
-class Character(BaseModel):
-    bot_name: str
-    bot_info: str
-    user_name: str
-    user_info: str
-    chat_history: list[dict]
-
+from app.model import Character
 
 client = MongoClient('mongodb://localhost:27017/', port=27017)
 db = client["CharacterAI"]
@@ -16,7 +7,7 @@ collection = db["character_info"]
 
 
 async def query_character_info_all(bot_name: str):
-    info_list = collection.find({'bot_name': bot_name})[0]
+    info_list = collection.find_one({'bot_name': bot_name})
     # 解包字典, 生成Character对象
     character_info = Character(**info_list)
     return character_info
@@ -39,13 +30,9 @@ async def create_character_info(character_info: Character):
     return True
 
 
-async def storage_chat_history(role: str, character_info: Character):
-    if role == 'assistant':
-        result = collection.update_one({'bot_name': character_info.bot_name},
-                                       {'$push': {'chat_history': character_info.chat_history}})
-    else:
-        result = collection.update_one({'bot_name': character_info.bot_name},
-                                       {'$push': {'chat_history': character_info.chat_history}})
+async def storage_chat_history(character_info: Character):
+    result = collection.update_one({'bot_name': character_info.bot_name},
+                                   {'$set': {'chat_history': character_info.dump_chat_history()}})
     if result.matched_count == 0:
         return False
     return True
