+++ app/main.py
@@ -1,6 +1,7 @@
 from fastapi import FastAPI
 from app.routers import character
 from fastapi.middleware.cors import CORSMiddleware
+import uvicorn
 
 app = FastAPI()
 app.include_router(character.router)
@@ -13,3 +14,7 @@ app.add_middleware(
     allow_methods=["*"],
     allow_headers=["*"],
 )
+
+if __name__ == "__main__":
+
+    uvicorn.run(app)
