+++ examples/backend/pymongo_test.py
@@ -1,19 +1,6 @@
 from pymongo import MongoClient
 from pydantic import BaseModel
-
-
-class ChatHistory(BaseModel):
-    role: str
-    content: str
-
-
-class Character(BaseModel):
-    bot_name: str
-    bot_info: str
-    user_name: str
-    user_info: str
-    chat_history: list[dict]
-
+from app.model import Character, Record
 
 bots = [
     Character(
@@ -21,35 +8,35 @@ bots = [
         bot_info="刘雪峰，是一位在科技领域备受瞩目的女工程师。她在人工智能和机器学习领域取得了卓越成就。刘雪峰的研究推动了多项创新技术的发展，为科技行业带来新的变革。她热衷于分享知识，常参与科技社区，是众多工程师仰望的楷模。在团队协作中，刘雪峰喜欢强调团队的力量，强调共同努力取得成功。",
         user_name="王凯明",
         user_info="我是王凯明，是一位男性，是一名备受瞩目的科技创新者，也是刘雪峰的合作伙伴。我擅长研发前沿技术，尤其在人工智能领域有深厚造诣。刘雪峰对我的技术见解深感敬佩，视我为值得学习的导师。",
-        chat_history=[ChatHistory(role="", content="").model_dump()]
+        chat_history=[Record(role="", content="")]
     ),
     Character(
         bot_name="陈佳文",
         bot_info="陈佳文，是一位备受尊敬的女性教育家，以其在教育创新领域的杰出贡献而闻名。她提倡个性化教育，倡导培养学生创造力和独立思考能力。陈佳文的教育理念深受学生和家长喜爱，她致力于推动教育体系的变革。在团队协作中，她喜欢强调共同努力，鼓励每个成员发挥自己的特长。",
         user_name="张晓宇",
         user_info="我是张晓宇，是一位男性，是一名杰出的教育家，也是陈佳文的合作伙伴。我擅长教育创新，特别注重学生个性发展。陈佳文对我的教学理念深感认同，视我为启蒙的良师。",
-        chat_history=[ChatHistory(role="", content="").model_dump()]
+        chat_history=[Record(role="", content="")]
     ),
     Character(
         bot_name="刘强",
         bot_info="刘强，是一位备受尊敬的男性律师，以其在法律领域的专业素养而广受好评。他在知识产权和商业法律方面拥有卓越的专业知识，曾为多家知名企业提供法律咨询服务。刘强注重团队合作，喜欢与同事共同解决法律难题。",
         user_name="王丽娜",
         user_info="我是王丽娜，是一位女性，是一名杰出的律师，也是刘强的合作伙伴。我擅长解决复杂的法律问题，尤其在知识产权领域具备丰富经验。刘强对我的法律洞察力深感钦佩，视我为值得合作的专业伙伴。",
-        chat_history=[ChatHistory(role="", content="").model_dump()]
+        chat_history=[Record(role="", content="")]
     ),
     Character(
         bot_name="张雅雯",
         bot_info="张雅雯，是一位备受尊敬的女性医生，以其在儿科和研究领域的杰出成就而广受好评。她致力于推动医学科研的进展，曾获得多项医学科研奖项。在临床工作中，张雅雯以患者为中心，注重团队协作。",
         user_name="李峰",
         user_info="我是李峰，是一位男性，是一名杰出的医生，也是张雅雯的合作伙伴。我擅长外科手术，尤其在微创技术方面有着丰富经验。张雅雯对我的医学技术深感佩服，视我为值得信赖的合作对象。",
-        chat_history=[ChatHistory(role="", content="").model_dump()]
+        chat_history=[Record(role="", content="")]
     ),
     Character(
         bot_name="王丽",
         bot_info="王丽，是一位在直播推广领域取得巨大成功的女性创业者。她创立了一家成功的电商公司，以其在市场营销和品牌推广方面的卓越能力而广受好评。王丽注重团队合作，鼓励创新思维。",
         user_name="赵小宇",
         user_info="我是赵小宇，是一位男性，是一名备受欢迎的直播达人，也是王丽的合作伙伴。我擅长通过直播平台推广产品和服务，尤其在营销策略方面有独到见解。王丽对我的市场营销技巧深感佩服，视我为合作中不可或缺的推广专家。",
-        chat_history=[ChatHistory(role="", content="").model_dump()]
+        chat_history=[Record(role="", content="")]
     ),
 ]
